+++ CoreHistogramView/src/au/gov/asd/tac/constellation/views/histogram/HistogramFilterOnSelectionPlugin.java
@@ -28,6 +28,7 @@ import au.gov.asd.tac.constellation.plugins.parameters.PluginParameter;
 import au.gov.asd.tac.constellation.plugins.parameters.PluginParameters;
 import au.gov.asd.tac.constellation.plugins.parameters.types.ElementTypeParameterValue;
 import au.gov.asd.tac.constellation.plugins.parameters.types.SingleChoiceParameterType;
+import au.gov.asd.tac.constellation.plugins.templates.PluginTags;
 import au.gov.asd.tac.constellation.plugins.templates.SimpleEditPlugin;
 import java.util.ArrayList;
 import java.util.List;
@@ -43,7 +44,7 @@ import org.openide.util.lookup.ServiceProvider;
  */
 @ServiceProvider(service = Plugin.class)
 @NbBundle.Messages("HistogramFilterOnSelectionPlugin=Filter Histogram on Selection")
-@PluginInfo(pluginType = PluginType.UPDATE, tags = {"LOW LEVEL"})
+@PluginInfo(pluginType = PluginType.UPDATE, tags = {PluginTags.LOW_LEVEL})
 public class HistogramFilterOnSelectionPlugin extends SimpleEditPlugin {
 
     public static final String ELEMENT_TYPE_PARAMETER_ID = PluginParameter.buildId(HistogramFilterOnSelectionPlugin.class, "element_type");