+++ CoreInteractiveGraph/src/au/gov/asd/tac/constellation/graph/interaction/plugins/clipboard/PasteGraphPlugin.java
@@ -29,6 +29,7 @@ import au.gov.asd.tac.constellation.plugins.parameters.PluginParameter;
 import au.gov.asd.tac.constellation.plugins.parameters.PluginParameters;
 import au.gov.asd.tac.constellation.plugins.parameters.types.ObjectParameterType;
 import au.gov.asd.tac.constellation.plugins.parameters.types.ObjectParameterType.ObjectParameterValue;
+import au.gov.asd.tac.constellation.plugins.templates.PluginTags;
 import au.gov.asd.tac.constellation.plugins.templates.SimpleEditPlugin;
 import org.openide.util.NbBundle.Messages;
 import org.openide.util.lookup.ServiceProvider;
@@ -48,7 +49,7 @@ import org.openide.util.lookup.ServiceProvider;
     "MSG_VxAttrs=Node attributes don't match",
     "MSG_TxAttrs=Transaction attributes don't match"
 })
-@PluginInfo(pluginType = PluginType.IMPORT, tags = {"IMPORT"})
+@PluginInfo(pluginType = PluginType.IMPORT, tags = {PluginTags.IMPORT})
 public final class PasteGraphPlugin extends SimpleEditPlugin {
 
     // The graph being pasted from the paste buffer.